data_storage/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Mateusz <01178563@pw.edu.pl>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = "^8.16.0"
kafka-python = "^2.0.2"
requests = "^2.32.3"
six = "^1.16.0"
kafka-python-ng = "^2.2.3"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
